+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
-### Added
+### Changed
+- Rename protected `AbstractRestController::getApiProblemResponse` to `apiProblem` and use it internally
 
 
 ## [0.2.0] - 2015-04-03